requirements.txt
aiogram==3.13.0
openai==1.35.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
alembic==1.13.1
cryptography==43.0.0
aiofiles==23.2.1
asyncpg==0.29.0
greenlet==3.4.0
httpx==0.27.2